requirements.txt
beautifulsoup4~=4.12.2
fake-useragent==1.2.1
Flask~=2.3.2
requests~=2.31.0
PyYAML~=6.0.1
numpy~=1.25.2
opencv-contrib-python==4.8.0.76